% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extensions.R, R/use_dictionary.R
\name{extensions}
\alias{extensions}
\alias{use_dictionary.tilt_profile}
\title{Extensions}
\usage{
\method{use_dictionary}{tilt_profile}(data, ...)
}
\arguments{
\item{data}{A '\link[tiltIndicator:tilt_profile]{tilt_profile}'.}

\item{...}{
  Arguments passed on to \code{\link[rsetools:use_dictionary]{rsetools::use_dictionary}}
  \describe{
    \item{\code{}}{}
  }}
}
\description{
Methods extending objects defined in other packages.

\itemize{
\item \code{\link[rsetools:use_dictionary]{rsetools::use_dictionary()}}.
}
}
\examples{
library(tiltIndicator)

profile <- tilt_profile(nest_levels(
  product = tibble(companies_id = "a", x = 1),
  company = tibble(companies_id = "a", y = TRUE)
))
class(profile)

profile

use_dictionary(profile)
}
