% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_packages.R
\name{match_packages}
\alias{match_packages}
\title{Help find all installed packages matching a pattern}
\usage{
match_packages(x)
}
\value{
Character vector.
}
\description{
It's useful when working with a family of packages which name share a pattern.
}
\examples{
# All packages starting with tilt
match_packages("^tilt")
}
