% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiltDevTools-package.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{use_dictionary}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{rsetools}{\code{\link[rsetools]{use_dictionary}}}
}}

