% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiltDevTools-package.R
\docType{package}
\name{tiltDevTools-package}
\alias{tiltDevTools}
\alias{tiltDevTools-package}
\title{tiltDevTools: Tools to Make Developing 'TILT' Packages Easier}
\description{
Tools to make developing 'tilt' packages easier.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/2DegreesInvesting/tiltDevTools}
  \item \url{https://2degreesinvesting.github.io/tiltDevTools/}
  \item Report bugs at \url{https://github.com/2DegreesInvesting/tiltDevTools/issues}
}

}
\author{
\strong{Maintainer}: Mauro Lepore \email{maurolepore@gmail.com} (\href{https://orcid.org/0000-0002-1986-7988}{ORCID})

}
\keyword{internal}
