% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_dictionary.R
\name{use_dictionary.tilt_profile}
\alias{use_dictionary.tilt_profile}
\title{Extend \code{use_dictionary()} for objects of class 'tilt_profile'}
\usage{
\method{use_dictionary}{tilt_profile}(data, ...)
}
\arguments{
\item{data}{A data frame or list of data frames.}

\item{...}{Unused but necessary to support for extensions.}
}
\value{
A \link[tibble:tibble]{tibble::tibble}.
}
\description{
Create a dictionary
}
\examples{
library(tiltIndicator)
library(rsetools)

# Fake output of a tilt indicator profile
product <- data.frame(companies_id = "a", product_col = 1)
company <- data.frame(companies_id = "a", company_col = 2)
nested <- nest_levels(product, company)
nested

nested |> class()
# Not what you want
nested |> use_dictionary()

profile <- nested |> tilt_profile()
profile |> class()
# Yes what you want
profile |> use_dictionary()
}
\seealso{
\code{\link[rsetools:use_dictionary]{rsetools::use_dictionary()}}.
}
