% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_github_workflow.R
\name{use_github_workflow}
\alias{use_github_workflow}
\alias{use_github_workflow_check_revdep}
\alias{use_github_workflow_check}
\title{Use a GitHub Actions workflow to check a reverse dependency}
\usage{
use_github_workflow_check_revdep(package)

use_github_workflow_check()
}
\arguments{
\item{package}{Character. The name of a package which is depends on the
package where this workflow runs -- i.e. it's a reverse dependency.}
}
\value{
The first argument, invisibly. Called for its side effect.
}
\description{
Use a GitHub Actions workflow to check a reverse dependency
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# A workflow to run R CMD check on your package
use_github_workflow_check()

# A workflow to run R CMD check on reverse dependencies of your package
use_github_workflow_check_revdep("tiltWebTool")
\dontshow{\}) # examplesIf}
}
