% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extensions-use_dictionary.R, R/extensions.R
\name{use_dictionary.tilt_profile}
\alias{use_dictionary.tilt_profile}
\alias{extensions}
\title{Extensions}
\usage{
\method{use_dictionary}{tilt_profile}(data, ...)
}
\arguments{
\item{data}{A '\link[tiltIndicator:tilt_profile]{tilt_profile}'.}

\item{...}{
  Arguments passed on to \code{\link[rsetools:use_dictionary]{rsetools::use_dictionary}}
  \describe{
    \item{\code{}}{}
  }}
}
\description{
\itemize{
\item \code{\link[rsetools:use_dictionary]{rsetools::use_dictionary()}}.
}

Methods extending objects defined in other packages.
}
\examples{
library(tiltIndicator)

profile <- tilt_profile(nest_levels(
  product = tibble(companies_id = "a", x = 1),
  company = tibble(companies_id = "a", y = TRUE)
))
class(profile)

profile

use_dictionary(profile)
}
