% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_dictionary.R
\name{extensions}
\alias{extensions}
\title{Extensions}
\description{
\itemize{
\item \code{\link[rsetools:use_dictionary]{rsetools::use_dictionary()}} is extended with a method for
\link[tiltIndicator:tilt_profile]{tiltIndicator::tilt_profile}.
}
}
\examples{
library(tiltIndicator)

# Fake output of a tilt indicator profile
product <- data.frame(companies_id = "a", product_col = 1)
company <- data.frame(companies_id = "a", company_col = 2)
nested <- nest_levels(product, company)
nested

nested |> class()

# Not very useful
nested |> use_dictionary()

profile <- nested |> tilt_profile()
profile |> class()

# Useful
# Yes what you want
profile |> use_dictionary()
}
