% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_github_workflow.R
\name{use_github_workflow}
\alias{use_github_workflow}
\alias{use_github_workflow_check_revdep}
\alias{use_github_workflow_check}
\title{Use GitHub-Actions workflows to check a package or a reverse dependency}
\usage{
use_github_workflow_check_revdep(package)

use_github_workflow_check()
}
\arguments{
\item{package}{Character. The name of the reverse dependency you want to
check. it must be a package under the organization 2DegreesInvesting.}
}
\value{
Called for its side effect.
}
\description{
These functions call \href{https://docs.github.com/en/actions/using-workflows/reusing-workflows#creating-a-reusable-workflow}{reusable workflows}
centrally maintained at 2DegreesInvesting/tiltDevTools.
\itemize{
\item \code{use_github_workflow_check()} calls the same workflow you can get
with \code{usethis::use_github_actions("check-standard")}. The benefit is that
it's standardized across tilt* repos and you don't have to update it
yourself.
\item \code{use_github_workflow_check_revdep()} works for packages and reverse
dependencies under the same, 2DegreesInvesting organization.
}
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Run R CMD check on your package
use_github_workflow_check()

# R CMD check on a reverse dependencies of your package
use_github_workflow_check_revdep("tiltWebTool")
\dontshow{\}) # examplesIf}
}
